[UX]: select files in explorer
