+++ src/components/sources-list-view/models.ts
@@ -1,0 +1,5 @@
+export interface TreeItem {
+  value: string;
+  label: string;
+  children?: TreeItem[];
+}
